Move TitleBar to a separate file & re-make styles
