+++ tgui/packages/tgui/layouts/Window.tsx
@@ -12,14 +12,13 @@ import {
   useLayoutEffect,
   useState,
 } from 'react';
-import { Box, Icon } from 'tgui-core/components';
-import { UI_DISABLED, UI_INTERACTIVE, UI_UPDATE } from 'tgui-core/constants';
+import { Box } from 'tgui-core/components';
+import { UI_DISABLED, UI_INTERACTIVE } from 'tgui-core/constants';
 import { classes } from 'tgui-core/react';
-import { decodeHtmlEntities, toTitleCase } from 'tgui-core/string';
+import { decodeHtmlEntities } from 'tgui-core/string';
 
 import { backendSuspendStart, globalStore, useBackend } from '../backend';
 import { useDebug } from '../debug';
-import { toggleKitchenSink } from '../debug/actions';
 import {
   dragStartHandler,
   recallWindowGeometry,
@@ -28,9 +27,9 @@ import {
 } from '../drag';
 import { createLogger } from '../logging';
 import { Layout } from './Layout';
+import { TitleBar } from './TitleBar';
 
 const logger = createLogger('Window');
-
 const DEFAULT_SIZE: [number, number] = [400, 600];
 
 type Props = Partial<{
@@ -113,7 +112,6 @@ export const Window = (props: Props) => {
   return suspended ? null : (
     <Layout className="Window" theme={theme}>
       <TitleBar
-        className="Window__titleBar"
         title={title || decodeHtmlEntities(config.title)}
         status={config.status}
         fancy={fancy}
@@ -175,81 +173,3 @@ const WindowContent = (props: ContentProps) => {
 };
 
 Window.Content = WindowContent;
-
-const statusToColor = (status) => {
-  switch (status) {
-    case UI_INTERACTIVE:
-      return 'good';
-    case UI_UPDATE:
-      return 'average';
-    case UI_DISABLED:
-    default:
-      return 'bad';
-  }
-};
-
-type TitleBarProps = Partial<{
-  canClose: boolean;
-  className: string;
-  fancy: boolean;
-  onClose: (e) => void;
-  onDragStart: (e) => void;
-  status: number;
-  title: string;
-}> &
-  PropsWithChildren;
-
-const TitleBar = (props: TitleBarProps) => {
-  const {
-    className,
-    title,
-    status,
-    canClose,
-    fancy,
-    onDragStart,
-    onClose,
-    children,
-  } = props;
-  const dispatch = globalStore.dispatch;
-
-  const finalTitle =
-    (typeof title === 'string' &&
-      title === title.toLowerCase() &&
-      toTitleCase(title)) ||
-    title;
-
-  return (
-    <div className={classes(['TitleBar', className])}>
-      {(status === undefined && (
-        <Icon className="TitleBar__statusIcon" name="tools" opacity={0.5} />
-      )) || (
-        <Icon
-          className="TitleBar__statusIcon"
-          color={statusToColor(status)}
-          name="eye"
-        />
-      )}
-      <div
-        className="TitleBar__dragZone"
-        onMouseDown={(e) => fancy && onDragStart && onDragStart(e)}
-      />
-      <div className="TitleBar__title">
-        {finalTitle}
-        {!!children && <div className="TitleBar__buttons">{children}</div>}
-      </div>
-      {process.env.NODE_ENV !== 'production' && (
-        <div
-          className="TitleBar__devBuildIndicator"
-          onClick={() => dispatch(toggleKitchenSink())}
-        >
-          <Icon name="bug" />
-        </div>
-      )}
-      {Boolean(fancy && canClose) && (
-        <div className="TitleBar__close TitleBar__clickable" onClick={onClose}>
-          ×
-        </div>
-      )}
-    </div>
-  );
-};
