+++ tgui/packages/tgui/styles/layouts/Window.scss
@@ -22,16 +22,6 @@
   );
 }
 
-.Window__titleBar {
-  position: fixed;
-  z-index: 1;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 32px;
-  height: base.rem(32px);
-}
-
 // Everything after the title bar
 .Window__rest {
   position: fixed;
